Add local install preset, PostsolveStorage getters, and use size_t for C API sizes (#28)
This PR introduces several improvements to the C API and build system:

## 1. Local Installation Preset
Adds a CMake preset for local installation to the `local_install`
directory with a corresponding task for streamlined build and install
processes. Updates `.gitignore` to exclude the `local_install`
directory.

## 2. PostsolveStorage Reduction Log Getters
Exposes the internal reduction log stored in `PostsolveStorage` through
new C API functions:
- `libpapilo_postsolve_storage_get_types()` - Get reduction types array
- `libpapilo_postsolve_storage_get_indices()` - Get indices array
- `libpapilo_postsolve_storage_get_values()` - Get values array  
- `libpapilo_postsolve_storage_get_start()` - Get start offsets array

These getters allow external implementations to access the complete
reduction log for custom postsolve implementations. The API includes
comprehensive documentation explaining the reduction log structure and
how to interpret each reduction type.

## 3. Use `size_t` Instead of `int` for Size-Related C API Functions
Replaced `int` with `size_t` for all size and count-related parameters
and return values to match C++ semantics and avoid unnecessary casting.

### Changes:
- **Size output parameters**: Changed from `int*` to `size_t*` in
functions like:
  - `libpapilo_problem_get_objective_coefficients()`
  - `libpapilo_problem_get_lower_bounds()`
  - `libpapilo_postsolve_storage_get_types()`
- And all other functions that return array sizes via output parameters

- **Count/size return values**: Changed from `int` to `size_t` in
functions like:
  - `libpapilo_reductions_get_size()`
  - `libpapilo_reductions_get_num_transactions()`
  - `libpapilo_statistics_get_nrounds()`
  - `libpapilo_statistics_get_num_presolvers()`
  - All per-presolver statistics functions (ncalls, nsuccessful, etc.)
  - `libpapilo_problem_update_get_singleton_cols_count()`

- **Error handling**: Replaced error-based returns (`-1`) with
`custom_assert` for validation, making the API more consistent with
modern C practices

### Benefits:
- Eliminates unnecessary `static_cast<int>()` conversions
- Matches C++ `size()` semantics directly
- Prevents potential narrowing issues with large data sets
- More idiomatic C API design

## 4. Documentation Fixes
Minor documentation fix: clarifies that "ntsx" means transactions, not
transformations.

---

All tests pass (33/33). ✓

---------

+++ test/libpapilo/GetterAPIsTest.cpp
@@ -70,12 +70,13 @@ TEST_CASE( "Reductions getters", "[libpapilo]" )
       REQUIRE( libpapilo_reductions_get_num_transactions( reductions ) == 1 );
 
       // Test transaction getters
-      int transaction_start =
+      size_t transaction_start =
           libpapilo_reductions_get_transaction_start( reductions, 0 );
-      int transaction_end =
+      size_t transaction_end =
           libpapilo_reductions_get_transaction_end( reductions, 0 );
-      int nlocks = libpapilo_reductions_get_transaction_nlocks( reductions, 0 );
-      int naddcoeffs =
+      size_t nlocks =
+          libpapilo_reductions_get_transaction_nlocks( reductions, 0 );
+      size_t naddcoeffs =
           libpapilo_reductions_get_transaction_naddcoeffs( reductions, 0 );
 
       REQUIRE( transaction_start == 0 );
@@ -136,7 +137,7 @@ TEST_CASE( "PostsolveStorage getters", "[libpapilo]" )
                1 );
 
       // Test mapping getters
-      int col_size = 0;
+      size_t col_size = 0;
       const int* col_mapping = libpapilo_postsolve_storage_get_orig_col_mapping(
           postsolve, &col_size );
       REQUIRE( col_mapping != nullptr );
@@ -144,7 +145,7 @@ TEST_CASE( "PostsolveStorage getters", "[libpapilo]" )
       REQUIRE( col_mapping[0] == 0 );
       REQUIRE( col_mapping[1] == 1 );
 
-      int row_size = 0;
+      size_t row_size = 0;
       const int* row_mapping = libpapilo_postsolve_storage_get_orig_row_mapping(
           postsolve, &row_size );
       REQUIRE( row_mapping != nullptr );
